requirements.txt
psycopg== 3.1.9
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
Flask==2.2.2